+++ lib/config.dart
@@ -6,22 +6,20 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:queuetodo/task.dart';
 
-void configSave(Queue<Task> tasks, int tasksMax) async {
+void configSave(Queue<Task> tasks) async {
   final preferences = await SharedPreferences.getInstance();
   final List<String> encodedTasks = [];
 
   for (Task taskObject in tasks) {
-    if (encodedTasks.length < tasksMax) {
-      try {
-        encodedTasks.add(json.encode({
-          'creationTimestamp': taskObject.creationTimestamp.toString(),
-          'lastModified': taskObject.lastModified.toString(),
-          'title': taskObject.titleController.text,
-          'description': taskObject.descriptionController.text,
-        }));
-      } on JsonUnsupportedObjectError {
-        continue;
-      }
+    try {
+      encodedTasks.add(json.encode({
+        'creationTimestamp': taskObject.creationTimestamp.toString(),
+        'lastModified': taskObject.lastModified.toString(),
+        'title': taskObject.titleController.text,
+        'description': taskObject.descriptionController.text,
+      }));
+    } on JsonUnsupportedObjectError {
+      continue;
     }
   }
   preferences.setStringList('_encodedTasks', encodedTasks);
@@ -35,9 +33,8 @@ Future<Queue<Task>> configRead(
   Queue<Task> tasks = Queue();
 
   for (String encoded in encodedTasks) {
-    final now = DateTime.now();
-
     if (tasks.length < tasksMax) {
+      final now = DateTime.now();
       try {
         decoded = json.decode(encoded);
       } on FormatException {