+++ lib/main.dart
@@ -4,6 +4,7 @@ import 'package:queuetodo/app.dart';
 import 'package:queuetodo/theme.dart';
 
 void main() {
+  WidgetsFlutterBinding.ensureInitialized();
   runApp(
     MaterialApp(
       theme: theme,