+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 ## 1.1.0
-- Task history.
+- Task archive.
 - Text length counters.
 - Multiline description.
+- A new icon.
 
 ## 1.0.0
 - Initial stable release - basic task queue.