+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 1.2.0
+- Polish localization.
+- More meaningful 'Usage' and 'About' tabs.
+
 ## 1.1.0
 - Text length counters.
 - Possible multiline description of a task.